+++ index.js
@@ -26,7 +26,7 @@ module.exports = {
   rules: {
     'prettier/prettier': 'error',
     'valid-jsdoc': ['error'],
-    'backpack/use-tokens': 'error',
+    'backpack/use-tokens': 'off',
     'backpack/use-components': 'error',
 
     // This rule is prettry restrictive and we feel this decision should be left to developers to decide on a case by case basis.